Update to match dev branch changes, fix several porting bugs, bump Kotlin and KTX versions

+++ isegoria/src/main/java/com/eulersbridge/isegoria/network/api/models/PhotoAlbum.kt
@@ -8,7 +8,7 @@ import kotlinx.android.parcel.Parcelize
 @SuppressLint("ParcelCreator")
 @Parcelize
 data class PhotoAlbum (
-    @Json(name = "nodeId") val id: Int = 0,
+    @Json(name = "nodeId") val id: Long = 0,
     val name: String? = null,
     val description: String?,
     val location: String?,
